Multy banks and timers (#34)
* Add multiTimer

+++ src/main/java/ua/goit/banks/CurrencyInterface.java
@@ -2,9 +2,9 @@ package ua.goit.banks;
 
 public interface CurrencyInterface {
 
-    double getCurrencySellingRate();
+    float getCurrencySellingRate();
 
-    double getCurrencyBuyingRate();
+    float getCurrencyBuyingRate();
 
     Currencies getName();
 }
